Improving sync between players
Working on improving sync between players. Previously the player who swung on the sandbag was not made the owner of the sandbag. This somehow seemed to work decently before, but some issues became apparent. Namely that the hit detection for the non owner player was done through the owners POV. This means that with the highly delayed and highly interpolated movement of the bat was used as hit detection, which often led to misses at higher swing speeds. Now the player who owns the bat is made the owner of the sandbag when they swing on it. This is standard as the player needs to own the Game Object in order to write to any UdonSycned variables that are in the Udon Behaviour. I will need to keep an eye on this later on as I can foresee some possible race conditions if multiple players swing on the sandbag at the same time. Perhaps a scheduling system could be implemented?