Working on Syncing Player's times
Committing the current code as of now because I am going to heavily change it. The idea is that I want to sync players to a repeating clock of 1 second. I figure in order to do this we cannot rely on the local system clock as provided by the OS as this is not 100% accurate and can be modified by the user. Instead, I want to use UdonSynced variables to send the time of the Instance Master to the players. Then by using some math I might be able to get the other player on to the same repeating cycle as the Instance Master.