Only instance owner calls the sync local time
Changed the code so that only the owner of the instance / network master calls the SyncLocalTime function. This then sets the local time variable to Time.realtimeSinceStartup. local time is a UdonSynced variable so we then Serialize this variable which pushes it to everyone. Then we can use the DeserializationResult class which contains information about the send and receive time of this data to calculate the offset we would need to use to synchronise times.