A somewhat working network sync between players
As of this code version, an int value is used to initiate a sync. This int has two properties on it: an UdonSynced property, and a FieldChangeCallback property. These are both properties unique to the Udon environment. The FieldChangeCallback essentially creates a higher layer above the variable it is attached two. Instead of writing to the variable, you write to the property. When the property is written to, it runs the code within the property. This effectively creates a OnVariableChange event.
When the Launch Sandbag button is pressed in game, the player who pressed the button becomes the owner of the Sandbag Game Object. This is needed to write to an UdonSynced variable. Then they assign a value to the property mentioned prior.
There is then code within the property which creates schedule delay for the Owner of the Sandbag. This is an attempt to reduce the effect of network delay between the two players.
As of this code version, bools[0] is the only condition considered. This currently corresponds to the Launch Sandbag state. I will likely need to reintroduce an UdonSynced variable that is an array of bools which should track which states we should process. The bools array was previously a UdonSynced variable but this was causing some small issues where the Owner of the Game Object was overwriting the status if they somehow processed theirs first. The eventual end goal for this system is for most of it to be simulated locally, with some UdonSynced variables assisting in syncing players together.