New script for locking player camera during swing
This commit contains a new script that aims to make the experience of playing on Desktop a better experience. One key issue I noticed is that swinging the bat at the sandbag is an unpleasant experience on desktop as you are required to quickly swing at the sandbag, which is unpleasant on desktop due to how fast the camera is moving. The script I am creating essentially 'locks' the players camera when the press and hold the use button on the bat, which should make the experience of swinging less nauseating.