New Explosive Upgrade
Added a new upgrade called Explosive that serves as a test for modifying the trajectory of the sandbag mid-flight. Initial tests seem to work well, however the scheduling system is still not 100% airtight so I am experiencing some mistiming between the two players in my tests. However it is rare for a mistiming to happen. Perhaps a system that periodically checks to see if players are where they should be is required.