Changing the Bat collision method to use Ray Casts instead, also changing the Player Camera system to a 3rd person system instead
This commit contains the beginning of two system changes. I want to change the method that the bat uses to detect collisions from a literal collider (which has some detection/accuracy problems at high speeds) to a Ray Cast which from my initial testing seems more capable of handing collisions at high speeds. One advantage as well of the ray cast is we get a more accurate real world position of the collision which will be useful for another system later on where I want to freeze the bat's position in place to simulate impact force in a 'hitstop' like system.
Also in this commit is the beginning of the code changes to make the player camera system a 3rd person system rather than simulating locking it to their 1st person.