Working on moving collision detection over to Ray Cast
Had to fix an issue where I wasn't getting the real world position of the child transform. This meant multiple ray casts were being drawn from the same point. This commit has it working for two of the batParts. Now I need to get all of the parts working, which I want to do with a for loop. Then I need to move all of the calculation code from the OnCollisionEnter into the FixedUpdate.