Working on floating damage text system and a 'hitstop' style system
This commit contains the work on a floating damage style system.  The general way it works is when a Raycast collision is detected it instantiates a prefab which has a text object in a canvas. This text object then plays an animation which moves the text upwards. When the animation is finished, it calls an animation event which destroys the object. This commit also has the work on a 'hitstop' style system, which aims to recreate the hitstop seen in fighting games when the bat collides with the sandbag. The animation for it is done through a script, but I also want the bat to temporarily remain stationary when a collision is detected, but I am still working on the maths required to determine where the bat should be.