Continuing work on the hit stop system
This commit contains the continued work on the hitstop system. I  am commit the current code but I believe through thorough debugging and testing that I know what the solution would be. For some reason I've been trying to calculate the hitstop position based on the hit.point returned from the collision. I could have just been using the transform of the bat and move it back a certain amount based on how far through the sandbag the bat is. I'll continue working on this tomorrow but that could be promising.