Improved hitstop style system
The code to calculate where the bat 'ghost' should be has been improved. The maths is now much more accurate (albeit some small improvements can be made) and the bat now mostly stops outside of the Sandbag. The code works by calculating the difference in velocities between the bat, and the bat part which makes contact with the Sandbag. Then it calculates how far the bat moved through the Sandbag. Then finally we moved the bat 'backwards' by a factor of this distance travelled through the Sandbag.