Working on adding a second weapon
This commit contains the current WIP for a second weapon. This will require some restructuring of the code, as all of the code for the weapon collision is in the Sandbag script. As of current, I imagine the code structure will have a script on each weapon. So I am moving the code from the sandbag into a new weapon for now, to test to make sure it works.