Adding a second weapon
As of this commit the second weapon is working. This commit also contains updated code so that certain events are networked to other players, like starting the bat shake animation. A new script is also added in this commit. The UpdateHasSwungState script can as the name suggests update the 'has swung' state on all of the weapons. This is primarily used to only allow one weapon to be used per person.